HHH-14848 - Move Ant tasks + support from hibernate-core into tooling/hibernate-ant
