+++ hibernate-core/src/main/java/org/hibernate/tool/schema/UniqueConstraintSchemaUpdateStrategy.java
@@ -1,10 +1,10 @@
 /*
  * Hibernate, Relational Persistence for Idiomatic Java
  *
- * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
- * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or http://www.gnu.org/licenses/lgpl-2.1.html
  */
-package org.hibernate.tool.hbm2ddl;
+package org.hibernate.tool.schema;
 
 import java.util.Locale;
 import org.jboss.logging.Logger;
