+++ hibernate-core/src/test/java/org/hibernate/orm/test/schemavalidation/NumericValidationTest.java
@@ -37,7 +37,6 @@ import org.hibernate.tool.schema.spi.SourceDescriptor;
 import org.hibernate.tool.schema.spi.TargetDescriptor;
 
 import org.hibernate.testing.TestForIssue;
-import org.hibernate.orm.test.legacy.S;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
