+++ hibernate-core/src/test/java/org/hibernate/orm/test/tool/schema/scripts/MultiLineImportWithTabsAndSpacesTest.java
@@ -10,7 +10,6 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.List;
 
-import org.hibernate.dialect.Dialect;
 import org.hibernate.tool.hbm2ddl.MultipleLinesSqlCommandExtractor;
 
 import org.hibernate.testing.TestForIssue;
