+++ hibernate-spatial/src/test/java/org/hibernate/spatial/integration/TestSpatialSchemaGeneration.java
@@ -11,7 +11,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.EnumSet;
 import java.util.List;
