+++ tooling/hibernate-ant/src/main/java/org/hibernate/tool/hbm2ddl/MultipleLinesSqlCommandExtractor.java
